Rendu thread-safe les wrappers sur les données de host-interfaces qui ne l'étaient pas.
git-svn-id: https://svn.etat-de-vaud.ch/SVN_SOCLE/registre/unireg/trunk@40052 cb3efa0a-b8b0-9f43-a4c8-343d8b7e4362

+++ unireg/business/src/main/java/ch/vd/uniregctb/interfaces/model/wrapper/LocaliteWrapper.java
@@ -35,7 +35,7 @@ public class LocaliteWrapper implements Localite {
 
 	public Integer getComplementNPA() {
 		final Integer c = target.getComplementNPA();
-		if (c == null || c.intValue() == 0) {	// un complément de 0 signifie pas de complément
+		if (c == null || c == 0) {	// un complément de 0 signifie pas de complément
 			return null;
 		}
 		else {
